gpuav: Create gpu_error_message file
